refactor(orchestration): split into driver / planner / resolve
The 2041-line packs/orchestration.rs had three composite themes
sharing one file: the outer execute() driver, a ~440-LOC planner
(plan_pack_inner + collect_pack_intents + build_gate_table +
diagnostic helpers), and pack-name resolution (resolve_pack_dir_name,
validate_pack_names). Tests followed the same fault lines: general
dispatcher tests, then a 420-LOC "Preprocessing integration tests"
section, then a 390-LOC "Template preprocessor integration tests"
section, then the resolve_pack_dir_name_* triplet sandwiched
between them.

Each theme now lives in its own file under orchestration/:

- mod.rs           — execute(), prepare_packs(), execute_intents,
                     run_handler_pipeline (the driver and runner glue)
                     + dispatcher tests; re-exports the planner +
                     resolve API so existing
                     `crate::packs::orchestration::*` paths keep
                     resolving
- planning.rs      — PackPlan, plan_pack/_inner, collect_pack_intents/
                     _with_preprocessors/_inner, build_gate_table,
                     filter_pre_preprocess_gates, missing_target_hints,
                     display_path_relative_to_home; preprocessing +
                     template integration tests live here too,
                     because they exercise the planner end-to-end
- resolve.rs       — resolve_pack_dir_name, validate_pack_names + the
                     three resolve_pack_dir_name_* tests
- test_support.rs  — Shared MockCommandRunner + make_context +
                     TestUpCommand (pub(super)) so all three test
                     suites use the same fixtures

No production change. Same 1140 tests still pass. Public API is
preserved via re-exports — every existing
`crate::packs::orchestration::X` path still resolves.
