+++ crates/dodot-lib/src/packs/orchestration/planning.rs
@@ -1,198 +1,24 @@
-//! Orchestration pipeline — the single entry point for executing
-//! commands across packs.
+//! Pack planner — the "what would we do" computation.
 //!
-//! `execute()` owns the outer loop: discover packs → load per-pack
-//! config → execute command → aggregate results.
+//! Owns [`plan_pack`] (the main planner), [`plan_pack_inner`] (the
+//! actual scan → preprocess → match-rules → group-by-handler →
+//! to_intents pipeline, ~210 LOC), the [`PackPlan`] result type,
+//! [`build_gate_table`] (`HostFacts`/`[gates]` merging), and the
+//! per-pack `collect_pack_intents` API plus its diagnostic helpers
+//! (`missing_target_hints`, `display_path_relative_to_home`).
 //!
-//! The shared [`ExecutionContext`] (every command's dependency bag) and
-//! the result types ([`PackResult`], [`ExecuteResult`]) plus the
-//! per-pack [`Command`] trait live in sibling modules
-//! ([`crate::packs::context`], [`crate::packs::types`]); they are
-//! re-exported here for the historical `crate::packs::orchestration::X`
-//! surface.
+//! The driver in `mod.rs` calls this layer to plan one pack at a time;
+//! the runner functions there then take the resulting intents and feed
+//! them to the executor.
 
 use tracing::{debug, info};
 
-use crate::execution::Executor;
 use crate::gates::{GateTable, HostFacts};
 use crate::handlers;
-use crate::operations::OperationResult;
-use crate::packs::{self, Pack};
+use crate::packs::context::ExecutionContext;
+use crate::packs::Pack;
 use crate::rules::{self, Scanner};
 use crate::Result;
-
-pub use crate::packs::context::ExecutionContext;
-pub use crate::packs::types::{Command, ExecuteResult, PackResult};
-
-// ── Pipeline ────────────────────────────────────────────────────
-
-/// Execute a command across all (or filtered) packs.
-///
-/// This is the single entry point for the orchestration pipeline:
-///
-/// 1. Load root config
-/// 2. Discover packs (filtering by name if specified)
-/// 3. For each pack: load merged config → execute command → collect result
-/// 4. Aggregate results
-pub fn execute(
-    command: &dyn Command,
-    pack_filter: Option<&[String]>,
-    ctx: &ExecutionContext,
-) -> Result<ExecuteResult> {
-    info!(command = command.name(), "starting command");
-
-    // Load root config for pack-level ignore patterns
-    let root_config = ctx.config_manager.root_config()?;
-    debug!(
-        ignore_patterns = ?root_config.pack.ignore,
-        "loaded root config"
-    );
-
-    // Discover packs
-    let mut all_packs = packs::discover_packs(
-        ctx.fs.as_ref(),
-        ctx.paths.dotfiles_root(),
-        &root_config.pack.ignore,
-    )?;
-    info!(
-        count = all_packs.len(),
-        root = %ctx.paths.dotfiles_root().display(),
-        "discovered packs"
-    );
-
-    // Validate and apply name filter
-    if let Some(names) = pack_filter {
-        let _warnings = validate_pack_names(names, ctx)?;
-        // Warnings are handled by the calling command (status/up/down)
-        debug!(filter = ?names, "applying pack filter");
-        all_packs.retain(|p| names.iter().any(|n| n == &p.display_name || n == &p.name));
-        info!(count = all_packs.len(), "packs after filter");
-    }
-
-    let total_packs = all_packs.len();
-    let mut pack_results = Vec::with_capacity(total_packs);
-    let mut successful = 0;
-    let mut failed = 0;
-    let host = ctx.host_facts.as_ref();
-
-    for mut pack in all_packs {
-        info!(pack = %pack.name, "processing pack");
-
-        // Load pack-specific merged config
-        let pack_config = match ctx.config_manager.config_for_pack(&pack.path) {
-            Ok(pack_config) => {
-                debug!(pack = %pack.name, "loaded pack config");
-                pack.config = pack_config.to_handler_config();
-                pack_config
-            }
-            Err(e) => {
-                info!(pack = %pack.name, error = %e, "pack config error, skipping");
-                failed += 1;
-                pack_results.push(PackResult {
-                    pack_name: pack.name.clone(),
-                    success: false,
-                    operations: Vec::new(),
-                    error: Some(format!("config error: {e}")),
-                });
-                continue;
-            }
-        };
-
-        // C3: skip packs gated out by `[pack] os` on this host. Counted
-        // as successful (it's the configured behaviour, not a failure)
-        // with no operations — same shape `.dodotignore` would have if
-        // it reached this loop.
-        if !crate::gates::pack_os_active(&pack_config.pack.os, host) {
-            debug!(
-                pack = %pack.name,
-                allowed = ?pack_config.pack.os,
-                current_os = %host.os,
-                "pack inactive on this OS, skipping"
-            );
-            successful += 1;
-            pack_results.push(PackResult {
-                pack_name: pack.name.clone(),
-                success: true,
-                operations: Vec::new(),
-                error: None,
-            });
-            continue;
-        }
-
-        match command.execute_for_pack(&pack, ctx) {
-            Ok(result) => {
-                if result.success {
-                    info!(pack = %pack.name, ops = result.operations.len(), "pack succeeded");
-                    successful += 1;
-                } else {
-                    info!(pack = %pack.name, ops = result.operations.len(), "pack completed with errors");
-                    failed += 1;
-                }
-                pack_results.push(result);
-            }
-            Err(e) => {
-                info!(pack = %pack.name, error = %e, "pack failed");
-                failed += 1;
-                pack_results.push(PackResult {
-                    pack_name: pack.name.clone(),
-                    success: false,
-                    operations: Vec::new(),
-                    error: Some(e.to_string()),
-                });
-            }
-        }
-    }
-
-    info!(
-        total = total_packs,
-        successful = successful,
-        failed = failed,
-        "command complete"
-    );
-
-    Ok(ExecuteResult {
-        pack_results,
-        total_packs,
-        successful_packs: successful,
-        failed_packs: failed,
-    })
-}
-
-// ── Pack preparation ────────────────────────────────────────────
-
-/// Discover, filter, and load config for all relevant packs.
-///
-/// Returns the list of packs ready for intent collection or command
-/// execution. This is the shared first step for commands that need
-/// to inspect multiple packs before acting (e.g. conflict detection).
-pub fn prepare_packs(pack_filter: Option<&[String]>, ctx: &ExecutionContext) -> Result<Vec<Pack>> {
-    let root_config = ctx.config_manager.root_config()?;
-
-    let mut all_packs = packs::discover_packs(
-        ctx.fs.as_ref(),
-        ctx.paths.dotfiles_root(),
-        &root_config.pack.ignore,
-    )?;
-    info!(count = all_packs.len(), "discovered packs");
-
-    if let Some(names) = pack_filter {
-        let _warnings = validate_pack_names(names, ctx)?;
-        debug!(filter = ?names, "applying pack filter");
-        all_packs.retain(|p| names.iter().any(|n| n == &p.display_name || n == &p.name));
-        info!(count = all_packs.len(), "packs after filter");
-    }
-
-    // Load per-pack config
-    let mut configured = Vec::with_capacity(all_packs.len());
-    for mut pack in all_packs {
-        let pack_config = ctx.config_manager.config_for_pack(&pack.path)?;
-        debug!(pack = %pack.name, "loaded pack config");
-        pack.config = pack_config.to_handler_config();
-        configured.push(pack);
-    }
-
-    Ok(configured)
-}
 
 // ── Built-in "up" pipeline helpers ──────────────────────────────
 
@@ -768,461 +594,24 @@ fn missing_target_hints(
         .collect()
 }
 
-/// Execute a pre-collected set of intents.
-///
-/// This is the second half of the two-phase execution model.
-/// Call [`collect_pack_intents`] first, run conflict detection,
-/// then call this to actually perform the mutations.
-pub fn execute_intents(
-    intents: Vec<crate::operations::HandlerIntent>,
-    ctx: &ExecutionContext,
-) -> Result<Vec<OperationResult>> {
-    let count = intents.len();
-    info!(
-        intents = count,
-        dry_run = ctx.dry_run,
-        force = ctx.force,
-        "executing intents"
-    );
-    let auto_chmod = ctx.config_manager.root_config()?.path.auto_chmod_exec;
-    let executor = Executor::new(
-        ctx.datastore.as_ref(),
-        ctx.fs.as_ref(),
-        ctx.paths.as_ref(),
-        ctx.dry_run,
-        ctx.force,
-        ctx.provision_rerun,
-        auto_chmod,
-    );
-    executor.execute(intents)
-}
-
-/// Run the standard handler pipeline for a pack: scan → match rules →
-/// group by handler → to_intents → execute.
-///
-/// Convenience wrapper that combines [`collect_pack_intents`] and
-/// [`execute_intents`]. Does **not** perform cross-pack conflict
-/// detection — use the two-phase API for that.
-pub fn run_handler_pipeline(pack: &Pack, ctx: &ExecutionContext) -> Result<Vec<OperationResult>> {
-    let intents = collect_pack_intents(pack, ctx)?;
-    execute_intents(intents, ctx)
-}
-
-/// Resolve a user-typed pack identifier to its on-disk directory
-/// name. Tries display name first, falls back to the raw on-disk
-/// name — so `dodot adopt nvim` and `dodot adopt 010-nvim` both find
-/// the same pack on disk.
-///
-/// Use this in commands that take a single pack-name argument and
-/// then need the directory path for filesystem or datastore work
-/// (`adopt`, `addignore`, `fill`). Errors with [`DodotError::PackNotFound`]
-/// when no match exists, and resolves through both active and ignored
-/// packs (the caller decides whether being ignored is fatal).
-pub fn resolve_pack_dir_name(input: &str, ctx: &ExecutionContext) -> crate::Result<String> {
-    let root_config = ctx.config_manager.root_config()?;
-    let scanned = packs::scan_packs(
-        ctx.fs.as_ref(),
-        ctx.paths.dotfiles_root(),
-        &root_config.pack.ignore,
-    )?;
-    if let Some(p) = scanned
-        .packs
-        .iter()
-        .find(|p| p.display_name == *input || p.name == *input)
-    {
-        return Ok(p.name.clone());
-    }
-    if let Some(dir) = scanned
-        .ignored
-        .iter()
-        .find(|d| d.as_str() == input || packs::display_name_for(d) == input)
-    {
-        return Ok(dir.clone());
-    }
-    Err(crate::DodotError::PackNotFound { name: input.into() })
-}
-
-/// Validate that requested pack names exist. Returns error for nonexistent
-/// packs and collects warnings for ignored packs.
-///
-/// Names resolve against the pack's *display name* (e.g. `nvim` for an
-/// on-disk `010-nvim`) first, then fall back to the raw on-disk name —
-/// so `dodot up nvim` and `dodot up 010-nvim` both find the same pack.
-/// The display name is the recommended form.
-pub fn validate_pack_names(names: &[String], ctx: &ExecutionContext) -> crate::Result<Vec<String>> {
-    let root_config = ctx.config_manager.root_config()?;
-    let scanned = packs::scan_packs(
-        ctx.fs.as_ref(),
-        ctx.paths.dotfiles_root(),
-        &root_config.pack.ignore,
-    )?;
-
-    let mut warnings = Vec::new();
-    for input in names {
-        if scanned
-            .packs
-            .iter()
-            .any(|p| p.display_name == *input || p.name == *input)
-        {
-            continue;
-        }
-        if scanned
-            .ignored
-            .iter()
-            .any(|dir| dir == input || packs::display_name_for(dir) == input)
-        {
-            warnings.push(format!("warning: pack '{}' is ignored, skipping", input));
-            continue;
-        }
-        return Err(crate::DodotError::PackNotFound {
-            name: input.clone(),
-        });
-    }
-    Ok(warnings)
-}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
+    #![allow(unused_imports)]
+
+    use std::sync::Arc;
+
+    use super::super::test_support::{make_context, MockCommandRunner, TestUpCommand};
+    use super::super::{
+        collect_pack_intents, execute, execute_intents, prepare_packs, run_handler_pipeline,
+    };
+    use super::{collect_pack_intents_with_preprocessors, plan_pack};
     use crate::config::ConfigManager;
-    use crate::datastore::{CommandOutput, CommandRunner, FilesystemDataStore};
+    use crate::datastore::CommandRunner;
+    use crate::datastore::FilesystemDataStore;
     use crate::fs::Fs;
+    use crate::packs::Pack;
     use crate::paths::Pather;
     use crate::testing::TempEnvironment;
-    use std::sync::{Arc, Mutex};
-
-    struct MockCommandRunner {
-        calls: Mutex<Vec<String>>,
-    }
-
-    impl MockCommandRunner {
-        fn new() -> Self {
-            Self {
-                calls: Mutex::new(Vec::new()),
-            }
-        }
-    }
-
-    impl CommandRunner for MockCommandRunner {
-        fn run(&self, executable: &str, arguments: &[String]) -> Result<CommandOutput> {
-            let cmd_str = format!("{} {}", executable, arguments.join(" "));
-            self.calls.lock().unwrap().push(cmd_str.trim().to_string());
-            Ok(CommandOutput {
-                exit_code: 0,
-                stdout: String::new(),
-                stderr: String::new(),
-            })
-        }
-    }
-
-    fn make_context(env: &TempEnvironment) -> ExecutionContext {
-        let runner: Arc<dyn crate::datastore::CommandRunner> = Arc::new(MockCommandRunner::new());
-        let datastore = Arc::new(FilesystemDataStore::new(
-            env.fs.clone(),
-            env.paths.clone(),
-            runner.clone(),
-        ));
-        let config_manager = Arc::new(ConfigManager::new(&env.dotfiles_root).unwrap());
-
-        ExecutionContext {
-            fs: env.fs.clone() as Arc<dyn Fs>,
-            datastore,
-            paths: env.paths.clone() as Arc<dyn Pather>,
-            config_manager,
-            syntax_checker: Arc::new(crate::shell::NoopSyntaxChecker),
-            command_runner: runner,
-            dry_run: false,
-            no_provision: true, // skip install/homebrew in tests
-            provision_rerun: false,
-            force: false,
-            view_mode: crate::commands::ViewMode::Full,
-            group_mode: crate::commands::GroupMode::Name,
-            verbose: false,
-            host_facts: Arc::new(crate::gates::HostFacts::detect()),
-        }
-    }
-
-    /// Simple command that runs the handler pipeline.
-    struct TestUpCommand;
-
-    impl Command for TestUpCommand {
-        fn name(&self) -> &str {
-            "test-up"
-        }
-
-        fn execute_for_pack(&self, pack: &Pack, ctx: &ExecutionContext) -> Result<PackResult> {
-            let operations = run_handler_pipeline(pack, ctx)?;
-            let success = operations.iter().all(|r| r.success);
-            // Mirror what the real up/down commands do: the user-facing
-            // pack identifier carried in `PackResult.pack_name` is the
-            // pack's display name, not its raw on-disk directory.
-            Ok(PackResult {
-                pack_name: pack.display_name.clone(),
-                success,
-                operations,
-                error: None,
-            })
-        }
-    }
-
-    #[test]
-    fn execute_discovers_and_processes_packs() {
-        let env = TempEnvironment::builder()
-            .pack("vim")
-            .file("vimrc", "set nocompatible")
-            .done()
-            .pack("git")
-            .file("gitconfig", "[user]\n  name = test")
-            .done()
-            .build();
-
-        let ctx = make_context(&env);
-        let result = execute(&TestUpCommand, None, &ctx).unwrap();
-
-        assert_eq!(result.total_packs, 2);
-        assert_eq!(result.successful_packs, 2);
-        assert_eq!(result.failed_packs, 0);
-        assert!(result.is_success());
-
-        // Both packs should have operations
-        for pr in &result.pack_results {
-            assert!(pr.success, "pack {} failed", pr.pack_name);
-            assert!(
-                !pr.operations.is_empty(),
-                "pack {} has no operations",
-                pr.pack_name
-            );
-        }
-    }
-
-    #[test]
-    fn execute_filters_by_pack_name() {
-        let env = TempEnvironment::builder()
-            .pack("vim")
-            .file("vimrc", "x")
-            .done()
-            .pack("git")
-            .file("gitconfig", "x")
-            .done()
-            .pack("zsh")
-            .file("zshrc", "x")
-            .done()
-            .build();
-
-        let ctx = make_context(&env);
-        let filter = vec!["vim".into(), "zsh".into()];
-        let result = execute(&TestUpCommand, Some(&filter), &ctx).unwrap();
-
-        assert_eq!(result.total_packs, 2);
-        let names: Vec<&str> = result
-            .pack_results
-            .iter()
-            .map(|r| r.pack_name.as_str())
-            .collect();
-        assert!(names.contains(&"vim"));
-        assert!(names.contains(&"zsh"));
-        assert!(!names.contains(&"git"));
-    }
-
-    #[test]
-    fn execute_filter_resolves_display_name_to_prefixed_pack() {
-        let env = TempEnvironment::builder()
-            .pack("010-brew")
-            .file("Brewfile", "x")
-            .done()
-            .pack("nvim")
-            .file("init.lua", "x")
-            .done()
-            .build();
-
-        let ctx = make_context(&env);
-        let filter = vec!["brew".into()];
-        let result = execute(&TestUpCommand, Some(&filter), &ctx).unwrap();
-
-        // Filter `brew` resolves to the on-disk `010-brew` pack via display name.
-        assert_eq!(result.total_packs, 1);
-        assert_eq!(result.pack_results[0].pack_name, "brew");
-    }
-
-    #[test]
-    fn execute_filter_accepts_raw_directory_name_as_fallback() {
-        let env = TempEnvironment::builder()
-            .pack("010-brew")
-            .file("Brewfile", "x")
-            .done()
-            .build();
-
-        let ctx = make_context(&env);
-        let filter = vec!["010-brew".into()];
-        let result = execute(&TestUpCommand, Some(&filter), &ctx).unwrap();
-
-        // The raw directory name is a valid fallback for muscle memory or scripts.
-        assert_eq!(result.total_packs, 1);
-        // PackResult.pack_name surfaces the display-name form regardless of how
-        // the user typed the filter — that's what every render path expects.
-        assert_eq!(result.pack_results[0].pack_name, "brew");
-    }
-
-    #[test]
-    fn resolve_pack_dir_name_finds_pack_by_display_name() {
-        let env = TempEnvironment::builder()
-            .pack("010-nvim")
-            .file("init.lua", "x")
-            .done()
-            .build();
-
-        let ctx = make_context(&env);
-        let resolved = resolve_pack_dir_name("nvim", &ctx).unwrap();
-        assert_eq!(resolved, "010-nvim");
-    }
-
-    #[test]
-    fn resolve_pack_dir_name_finds_pack_by_raw_directory_name() {
-        let env = TempEnvironment::builder()
-            .pack("010-nvim")
-            .file("init.lua", "x")
-            .done()
-            .build();
-
-        let ctx = make_context(&env);
-        let resolved = resolve_pack_dir_name("010-nvim", &ctx).unwrap();
-        assert_eq!(resolved, "010-nvim");
-    }
-
-    #[test]
-    fn resolve_pack_dir_name_errors_on_unknown_pack() {
-        let env = TempEnvironment::builder()
-            .pack("vim")
-            .file("vimrc", "x")
-            .done()
-            .build();
-
-        let ctx = make_context(&env);
-        let err = resolve_pack_dir_name("nope", &ctx).unwrap_err();
-        assert!(matches!(
-            err,
-            crate::DodotError::PackNotFound { ref name } if name == "nope"
-        ));
-    }
-
-    #[test]
-    fn execute_skips_dodotignored_packs() {
-        let env = TempEnvironment::builder()
-            .pack("vim")
-            .file("vimrc", "x")
-            .done()
-            .pack("disabled")
-            .file("stuff", "x")
-            .ignored()
-            .done()
-            .build();
-
-        let ctx = make_context(&env);
-        let result = execute(&TestUpCommand, None, &ctx).unwrap();
-
-        assert_eq!(result.total_packs, 1);
-        assert_eq!(result.pack_results[0].pack_name, "vim");
-    }
-
-    #[test]
-    fn run_handler_pipeline_creates_symlinks() {
-        let env = TempEnvironment::builder()
-            .pack("vim")
-            .file("vimrc", "set nocompatible")
-            .file("gvimrc", "set guifont=Mono")
-            .done()
-            .build();
-
-        let ctx = make_context(&env);
-        let pack = Pack::new(
-            "vim".into(),
-            env.dotfiles_root.join("vim"),
-            ctx.config_manager
-                .config_for_pack(&env.dotfiles_root.join("vim"))
-                .unwrap()
-                .to_handler_config(),
-        );
-
-        let results = run_handler_pipeline(&pack, &ctx).unwrap();
-        assert!(results.iter().all(|r| r.success));
-
-        // Verify symlinks were created
-        let vim_symlink_dir = ctx.paths.handler_data_dir("vim", "symlink");
-        assert!(ctx.fs.exists(&vim_symlink_dir));
-    }
-
-    #[test]
-    fn dry_run_produces_results_without_side_effects() {
-        let env = TempEnvironment::builder()
-            .pack("vim")
-            .file("vimrc", "x")
-            .done()
-            .build();
-
-        let runner: Arc<dyn crate::datastore::CommandRunner> = Arc::new(MockCommandRunner::new());
-        let datastore = Arc::new(FilesystemDataStore::new(
-            env.fs.clone(),
-            env.paths.clone(),
-            runner.clone(),
-        ));
-        let config_manager = Arc::new(ConfigManager::new(&env.dotfiles_root).unwrap());
-
-        let ctx = ExecutionContext {
-            fs: env.fs.clone() as Arc<dyn Fs>,
-            datastore,
-            paths: env.paths.clone() as Arc<dyn Pather>,
-            config_manager,
-            syntax_checker: Arc::new(crate::shell::NoopSyntaxChecker),
-            command_runner: runner,
-            dry_run: true,
-            no_provision: true,
-            provision_rerun: false,
-            force: false,
-            view_mode: crate::commands::ViewMode::Full,
-            group_mode: crate::commands::GroupMode::Name,
-            verbose: false,
-            host_facts: Arc::new(crate::gates::HostFacts::detect()),
-        };
-
-        let result = execute(&TestUpCommand, None, &ctx).unwrap();
-        assert!(result.is_success());
-        assert!(!result.pack_results[0].operations.is_empty());
-
-        // No filesystem changes should have been made
-        let vim_symlink_dir = ctx.paths.handler_data_dir("vim", "symlink");
-        assert!(!ctx.fs.exists(&vim_symlink_dir));
-    }
-
-    #[test]
-    fn no_provision_skips_install_handler() {
-        let env = TempEnvironment::builder()
-            .pack("vim")
-            .file("vimrc", "set nocompatible")
-            .file("install.sh", "#!/bin/sh\necho setup")
-            .done()
-            .build();
-
-        let ctx = make_context(&env); // no_provision = true
-
-        let pack = Pack::new(
-            "vim".into(),
-            env.dotfiles_root.join("vim"),
-            ctx.config_manager
-                .config_for_pack(&env.dotfiles_root.join("vim"))
-                .unwrap()
-                .to_handler_config(),
-        );
-
-        let results = run_handler_pipeline(&pack, &ctx).unwrap();
-
-        // Should have symlink operations but no RunCommand
-        for r in &results {
-            assert!(
-                !matches!(r.operation, crate::operations::Operation::RunCommand { .. }),
-                "RunCommand should be skipped with no_provision"
-            );
-        }
-    }
 
     // ── Preprocessing integration tests ────────────────────────
 
@@ -1641,9 +1030,9 @@ mod tests {
             _ => false,
         });
         assert!(
-            has_expanded_source,
-            "production collect_pack_intents should expand .tar.gz via the default registry. Intents: {intents:?}"
-        );
+        has_expanded_source,
+        "production collect_pack_intents should expand .tar.gz via the default registry. Intents: {intents:?}"
+    );
     }
 
     // ── Template preprocessor integration tests ─────────────────
